Upgraded NextJS to 14.2.8. Cleaned up project structure: Consolidated files and yup schema in @/app/config and removed the constants/ folder.
