+++ packages/app/config/pages.ts
