+++ packages/app/config/validation-schema.ts
@@ -1,5 +1,5 @@
-import { InferType, array, mixed, number, object, string } from 'yup';
-import { linkTypes, titles, validFileExtensions } from '.';
+import { InferType, array, boolean, mixed, number, object, string } from 'yup';
+import { linkTypes, titles, validFileExtensions } from './constants';
 
 export type Title = (typeof titles)[number];
 export type LinkTypes = (typeof linkTypes)[number];
@@ -130,3 +130,43 @@ export const groupFormSchema = object({
 });
 
 export type GroupFormType = InferType<typeof groupFormSchema>;
+
+export const profileFormSchema = object({
+  avatar: mixed()
+    .test('is-valid-type', 'Not a valid image type', (value: unknown) => {
+      if (value instanceof File) {
+        return isValidFileType(value && value.name?.toLowerCase(), 'image');
+      }
+      // TODO check if its a url
+      return true;
+    })
+    .test('is-valid-size', 'Maximum image size allowed is 3MB', (value: unknown) => {
+      if (value instanceof File) {
+        return value.size <= megabytesToBytes(3);
+      }
+      return true;
+    }),
+  title: string().nullable().oneOf(titles, 'Not a valid title'),
+  name: string().nullable(),
+  nickname: string().nullable(),
+  email: string().email('Not a valid email').required('Email is required'),
+  phone: string(),
+  links: linksSchema,
+});
+
+export type ProfileFormType = InferType<typeof profileFormSchema>;
+
+export const createNewGroupFormSchema = object({
+  name: string().required('Group name is required'),
+  verified: boolean().default(false),
+  city: string(),
+  country: string().required('Country is required'),
+});
+
+export type CreateNewGroupFormType = InferType<typeof createNewGroupFormSchema>;
+
+export const joinGroupFormSchema = object({
+  id: string().required('Please select a group'),
+});
+
+export type JoinGroupFormType = InferType<typeof joinGroupFormSchema>;
