+++ packages/app/package.json
@@ -60,13 +60,13 @@
     "intl-messageformat": "^10.5.14",
     "lodash": "^4.17.21",
     "lucide-react": "^0.436.0",
-    "next": "14.2.7",
+    "next": "14.2.8",
     "next-auth": "^4.24.7",
     "next-themes": "^0.3.0",
     "notistack": "^3.0.1",
     "nuqs": "^1.17.8",
     "pino-pretty": "^11.2.2",
-    "postcss": "8.4.41",
+    "postcss": "8.4.45",
     "postgres": "^3.4.4",
     "react": "18.3.1",
     "react-dom": "18.3.1",
