+++ packages/www/package.json
@@ -38,9 +38,9 @@
     "i18next-browser-languagedetector": "^8.0.0",
     "i18next-resources-to-backend": "^1.2.1",
     "intl-messageformat": "^10.5.14",
-    "next": "14.2.7",
+    "next": "14.2.8",
     "next-themes": "^0.3.0",
-    "postcss": "8.4.41",
+    "postcss": "8.4.45",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-i18next": "^15.0.1",
@@ -52,7 +52,7 @@
     "@typescript-eslint/eslint-plugin": "^8.3.0",
     "@typescript-eslint/parser": "^8.3.0",
     "chai": "^4.5.0",
-    "cssnano": "^7.0.5",
+    "cssnano": "^7.0.6",
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.7",
     "eslint-config-prettier": "^9.1.0",
